test: move s3 list encode regression to its own file
